Extended deploy test
